api/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
flask-restful
psycopg2-binary
pandas
